Revized theming

+++ lib/tasks/assets.rake
@@ -2,28 +2,45 @@ namespace :curriculr do
   namespace :assets do
     desc "Compile all the assets named in config.assets.precompile and runs css-flip to support RTL."
     task :precompile => :environment do
-      # Make sure you have css-flip installed
-      Dir.chdir Rails.root 
-      `npm install css-flip`
-
-      # Generate en assets
-      assets_path = "#{Rails.root}/public/assets"
-      `rm -rf #{Rails.root}/public/assets/*`
-      `rm -f #{Rails.root}/app/assets/stylesheets/sunshine/style_rtl.css`
-
-      Rake::Task["assets:precompile"].execute
-
-      # Identify and locate the en assets that will be css-flipped
-      Dir.chdir "#{assets_path}"
-      css_path =  %(#{assets_path}/#{Dir.glob("style-*.css").first})
-
-      # Css-flip the indentifed en assets
-      Dir.chdir Rails.root
-      puts "Runnig cssflip for theme: sunshine and flavor: #{ENV['THEME_FLAVOR'] || 'vanilla'}..."
-      `./bin/cssflip.sh . style #{css_path}`
+      base_path = "#{Rails.root}/vendor/assets/stylesheets/semantic_ui"
+      to_base_path = "#{base_path}/config"
+      `mkdir #{to_base_path}` unless Dir.exist?(to_base_path)
       
-      # Done
-      puts "RTL css generated successfully. Run rails assets:precompile to finish."
+      Account.all.each do |a|
+        theme = a.config['theme']['name']
+        
+        if Dir.exist?("#{base_path}/themes/#{theme}")
+          # Copy theme files for preprocessing
+          from_base_path = "#{base_path}/themes/#{theme}"
+          `cp -R #{base_path}/themes/config/* #{to_base_path}/`
+          `cp -R #{from_base_path}/* #{to_base_path}/`
+          
+          # Make sure you have css-flip installed
+          Dir.chdir Rails.root 
+          `npm install css-flip`
+
+          # Generate en assets
+          assets_path = "#{Rails.root}/public/assets"
+          `rm -rf #{Rails.root}/public/assets/*`
+          `rm -f #{Rails.root}/app/assets/stylesheets/sunshine/style_rtl.css`
+
+          Rake::Task["assets:precompile"].execute
+
+          # Identify and locate the en assets that will be css-flipped
+          Dir.chdir "#{assets_path}"
+          css_path =  %(#{assets_path}/#{Dir.glob("style-*.css").first})
+
+          # Css-flip the indentifed en assets
+          Dir.chdir Rails.root
+          puts "Runnig cssflip for theme: sunshine and flavor: #{ENV['THEME_FLAVOR'] || 'vanilla'}..."
+          `./bin/cssflip.sh . style #{css_path}`
+          
+          #Rake::Task["curriculr:assets:precompile"].execute
+          
+          # Done
+          puts "RTL css generated successfully. Run rails assets:precompile to finish."
+        end
+      end
     end
   end
 end
